requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
decorator==4.4.2
Django==4.1.13
dnspython==2.6.1
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
lazy_loader==0.3
moviepy==1.0.3
networkx==3.2.1
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pillow==10.2.0
proglog==0.1.10
pymongo==3.13.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
scikit-image==0.22.0
scipy==1.12.0
sqlparse==0.4.4
tifffile==2024.2.12
tqdm==4.66.2
typing_extensions==4.10.0
urllib3==2.2.1
